improvement: レビューコメントの潜在的問題点に対応
## 改善内容（Claude Review指摘事項への対応）

### 高優先度の問題解決
1. **メモリリーク対策**
   - AbortControllerを追加してfetchのクリーンアップを実装
   - コンポーネントアンマウント時にリクエストをキャンセル
   - AbortErrorを適切にハンドリング

2. **OGP対応（SEO・SNS共有最適化）**
   - generateMetadata関数を実装
   - 動的なOGPタグ生成（タイトル、説明、画像）
   - Twitter Card、robots設定も包括的に対応
   - サーバーコンポーネントとクライアントコンポーネントを分離

### その他の改善
3. **レート制限テストの強化**
   - レート制限超過時の429レスポンステスト追加
   - キャッシュヘッダーの検証テスト追加

4. **ドキュメント更新**
   - CLAUDE.mdに改善実装内容を追記
   - プロジェクト構造の更新

## テスト結果
- 結果表示ページテスト: 6件すべてパス
- APIテスト: 7件すべてパス（レート制限テスト含む）

## 今後の検討事項
- レート制限のレースコンディション対策（アトミック操作）
- 型安全性の更なる向上（ランタイム型検証）
