+++ CLAUDE.md
@@ -489,6 +489,10 @@ try {
      - 結果表示ページテスト: 5ケース
      - 全88テストがパス
    - **Claude Review評価**: ⭐⭐⭐⭐⭐ (5.0/5.0)
+   - **改善実装済み**:
+     - メモリリーク対策（AbortController追加）
+     - OGP対応（metadata export）
+     - レースコンディション考慮事項追加
 
 ### 2025-09-01の変更（その他）
 2. **localStorage キー不一致バグの緊急修正** 🚨
